+++ tomatic/callinfo.py
@@ -117,6 +117,9 @@ class CallInfo(object):
         return list(result)
 
     def partnerInfo(self, partner_data):
+        partner_category_id = self.O.IrModelData.get_object_reference(
+            'som_partner_account','res_partner_category_soci')[1]
+
         result = ns(
             id_soci=self.anonymize(partner_data.ref) if partner_data.ref else "",
             lang=partner_data.lang,
@@ -134,6 +137,7 @@ class CallInfo(object):
                 ([self.anonymize(partner_data.www_mobile)] if partner_data.www_mobile else [])
             ),
             comment=partner_data.comment or '',
+            is_member=partner_category_id in partner_data.category_id,
         )
         return result